pyproject.toml
[project]
name = "tickettracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-allauth[headless]>=65.13.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
]
